gensrclist.cc: implemented --prev-stdin option
This change explores the technique of reusing previous output as a
cache.  With gensrclist, headers are sorted according to *.src.rpm
filenames.  Hence a relatively simple approach is possible: just merge
headers from the cache with filenames which needs to be processed (in
other words, "rewind" the cache to the current filename).  This approach
probably won't be possible with genpkglist, where the headers are first
grouped by %{sourcerpm}.