genpkglist.cc: restored partial changelog copying
Legionus, a son of a gun and an illustrious moderator,
has conned everyone by simply not porting partial changelog
copying in 6a3fd0ff.  Instead, changelogs started to get copied
in full.  What a botched job!

With this change, the output is again byte-to-byte identical
to what they had before.

$ LD_LIBRARY_PATH=~tmp ~tmp/genpkglist --changelog-since 2147483647 \
	--cachedir ~/.cache --no-scan /klad/Sisyphus/x86_64 classic &&
	stat /klad/Sisyphus/x86_64/base/pkglist.classic |grep Mod &&
	md5sum /klad/Sisyphus/x86_64/base/pkglist.classic
Modify: 2017-03-23 02:57:07.827238752 +0300
11a4f78f46de33f9677193c9b3da10aa  /klad/Sisyphus/x86_64/base/pkglist.classic

$ ./genpkglist --changelog-since 2147483647 \
	--cachedir ~/.cache --no-scan /klad/Sisyphus/x86_64 classic &&
	stat /klad/Sisyphus/x86_64/base/pkglist.classic.lz4 |grep Mod &&
	lz4 -d </klad/Sisyphus/x86_64/base/pkglist.classic.lz4 |md5sum
Modify: 2017-03-23 02:57:29.061434908 +0300
11a4f78f46de33f9677193c9b3da10aa  -