zhdr.h: implemented unzhdrv()
A set of headers compressed in a single lz4 frame can now be
decompressed essentially without any extra information (such as the
number of headers and their sizes).  This requires only a little bit
of extra work.  First, the uncompressed size has to be stored in
in the frame (which is disabled by default).  Second, the size of
a header can be deduced rather easily.  To be extra sure, I assert
upon the leading magic.