genpkglist.cc: implemented --prev-stdin option
Unlike gensrclist, genpkglist cannot use a simple merge-like strategy
on stdin and dirEntries.  That's because genpkglist first groups (sorts)
packages by src.rpm before spewing them out.  Nonetheless, an optimal
strategy which preserves grouping is still possible: two adjacent
packages exposed on stdin can from a group (pertaining to the same
src.rpm package) if they also exist and are furthermore adjacent among
the dirEntries.

Note that only bloated pkglists can be reused as a cache.  In a
stripped-down pkglist, headers are interdependent with each other
(i.e. each header depends upon the list of useful files).

This change produces a byte-to-byte identical output.  In the
following example, I reuse p8 pkglist as a cache for Sisyphus pkglist.
The only restriction is that the same date should be used with the
--changelog-since option.  (In the postprocess() routine, file lists
are stripped, but changelogs are copied as is.  Another possibility
would be to pass genbasedir something like --changelog-since=tomorrow,
which should keep only the last changelog entry.)

$ ./genpkglist --bloat --changelog-since $(date -d 2009-01-01 +%s) /klad/p8/x86_64 classic

$ ./genpkglist         --changelog-since $(date -d 2009-01-01 +%s) /klad/Sisyphus/x86_64 classic
$ md5sum /klad/Sisyphus/x86_64/base/pkglist.classic.lz4
88c367135d3952e74329a75c8e457b43  /klad/Sisyphus/x86_64/base/pkglist.classic.lz4

$ lz4 -d </klad/p8/x86_64/base/pkglist.classic.lz4 |
  ./genpkglist --prev-stdin --changelog-since $(date -d 2009-01-01 +%s) /klad/Sisyphus/x86_64 classic
$ md5sum /klad/Sisyphus/x86_64/base/pkglist.classic.lz4
88c367135d3952e74329a75c8e457b43  /klad/Sisyphus/x86_64/base/pkglist.classic.lz4

The following example further ensures that the caching actually works.

$ strace -e trace=open ./genpkglist              --changelog-since $(date -d 2009-01-01 +%s) /klad/Sisyphus/x86_64 classic |&
  grep 'open(".*\.rpm"' |wc -l
42284

$ lz4 -d </klad/p8/x86_64/base/pkglist.classic.lz4 |
  strace -e trace=open ./genpkglist --prev-stdin --changelog-since $(date -d 2009-01-01 +%s) /klad/Sisyphus/x86_64 classic |&
  grep 'open(".*\.rpm"' |wc -l
18478